Rosbridge unity subscriber to compressed images
Mathias, I'm Francisco Bonin, a colleague of Miquel Massot, the guy how forked your project some months ago.  
We are trying to send images from ROS to Unity to be displayed in a scene. I have incorporated the "CompressedImageMsg.cs" to my project but I'm trying to find out how to develop the subscriber and how to connect it with an element of the scene. Have you any example of the whole thing that you could share with us ?